Tyuiu.NikiforovFA.Sprint0.Task5.V0.Lib/DataService.cs
namespace Tyuiu.NikiforovFA.Sprint0.Task5.V0.Lib
{
    public class DataService
    {
        public static int Addition(int a, int b)
        {
            return a + b;
        }
        public static int Subtraction(int a, int b)
        {
            return a - b;
        }
        public static int Multiplication(int a, int b)
        {
            return a * b;
        }
        public static int Division(int a, int b)
        {
            if (b != 0)
            {
                return a / b;
            }
            else
            {
                Console.WriteLine("Нельзя делить на 0");
                return -1;
            }
        }
    }
}